+++ backend/chainlit/__init__.py
@@ -1,5 +1,4 @@
 import os
-from datetime import datetime, timezone
 
 from dotenv import load_dotenv
 
@@ -102,16 +101,6 @@ def sleep(duration: int):
         duration (int): The duration in seconds.
     """
     return asyncio.sleep(duration)
-
-
-def utc_now():
-    dt = datetime.now(timezone.utc).replace(tzinfo=None)
-    return dt.isoformat() + "Z"
-
-
-def timestamp_utc(timestamp: float):
-    dt = datetime.fromtimestamp(timestamp, timezone.utc).replace(tzinfo=None)
-    return dt.isoformat() + "Z"
 
 
 @dataclass()
