+++ backend/chainlit/llama_index/callbacks.py
@@ -6,10 +6,10 @@ from llama_index.core.callbacks.schema import CBEventType, EventPayload
 from llama_index.core.llms import ChatMessage, ChatResponse, CompletionResponse
 from llama_index.core.tools.types import ToolMetadata
 
-from chainlit import utc_now
 from chainlit.context import context_var
 from chainlit.element import Text
 from chainlit.step import Step, StepType
+from chainlit.utils import utc_now
 
 DEFAULT_IGNORE = [
     CBEventType.CHUNKING,
