+++ backend/chainlit/step.py
@@ -10,7 +10,6 @@ from typing import Callable, Dict, List, Optional, TypedDict, Union
 from literalai import BaseGeneration
 from literalai.observability.step import StepType, TrueStepType
 
-from chainlit import utc_now
 from chainlit.config import config
 from chainlit.context import CL_RUN_NAMES, context, local_steps
 from chainlit.data import get_data_layer
@@ -18,6 +17,7 @@ from chainlit.element import Element
 from chainlit.logger import logger
 from chainlit.telemetry import trace_event
 from chainlit.types import FeedbackDict
+from chainlit.utils import utc_now
 
 
 def check_add_step_in_cot(step: "Step"):
