+++ backend/chainlit/utils.py
@@ -3,6 +3,7 @@ import importlib
 import inspect
 import os
 from asyncio import CancelledError
+from datetime import datetime, timezone
 from typing import Callable
 
 import click
@@ -15,6 +16,16 @@ from chainlit.auth import ensure_jwt_secret
 from chainlit.context import context
 from chainlit.logger import logger
 from chainlit.message import ErrorMessage
+
+
+def utc_now():
+    dt = datetime.now(timezone.utc).replace(tzinfo=None)
+    return dt.isoformat() + "Z"
+
+
+def timestamp_utc(timestamp: float):
+    dt = datetime.fromtimestamp(timestamp, timezone.utc).replace(tzinfo=None)
+    return dt.isoformat() + "Z"
 
 
 def wrap_user_function(user_function: Callable, with_task=False) -> Callable:
