+++ cypress/e2e/data_layer/main.py
@@ -4,7 +4,6 @@ from typing import Dict, List, Optional
 
 import chainlit as cl
 import chainlit.data as cl_data
-from chainlit import utc_now
 from chainlit.data.utils import queue_until_user_message
 from chainlit.element import Element, ElementDict
 from chainlit.socket import persist_user_session
@@ -17,6 +16,7 @@ from chainlit.types import (
     ThreadDict,
     ThreadFilter,
 )
+from chainlit.utils import utc_now
 
 now = utc_now()
